+++ docs/content/3.middleware/4.reference/multistore/change-log.md
@@ -1,5 +1,21 @@
 # Change log
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - @vue-storefront/middleware@3.9.0
+
+```diff [package.json]
+dependencies: {
+-  "@vue-storefront/middleware": "3.x.x",
+-  "@vue-storefront/multistore": "2.x.x",
++  "@vue-storefront/middleware": "^3.9.0",
++  "@vue-storefront/multistore": "^3.0.0",
+}
+```
+
 ## 2.0.1
 
 ### Patch Changes